split fp types to different files, add tf32
