+++ lib/src/arithmetic/signals/signals.dart
@@ -1,0 +1,5 @@
+// Copyright (C) 2023-2025 Intel Corporation
+// SPDX-License-Identifier: BSD-3-Clause
+
+export 'fixed_point_logic.dart';
+export 'floating_point_logics/floating_point_logics.dart';
